codeToImage/codeToImage/requirements.txt
cffi==1.16.0
numpy==1.26.4
pillow==10.2.0
pycparser==2.21
qrcode==7.3.1
webp==0.3.0